Share diff harness plumbing
Extract common IL/C# diff harness helpers for output format parsing, pair manifests, and path labels, and route shared helper changes through both targeted smoke jobs.
