+++ tools/CSharpDiffHarness/CSharpDiffHarness.csproj
@@ -13,6 +13,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <ProjectReference Include="../DiffHarnessCommon/DiffHarnessCommon.csproj" />
     <ProjectReference Include="../../src/ILInspector.Decompiler/ILInspector.Decompiler.csproj" />
     <PackageReference Include="Markout" />
   </ItemGroup>
